python-backend/requirements.txt
browser-use[memory]
fastapi
uvicorn[standard]
python-dotenv
pydantic
asyncio-mqtt
supabase
langchain-openai
langchain-anthropic
playwright 